face recognition f/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
